Move epoch.
