+++ astronomy/epoch_body.hpp
@@ -1,13 +1,10 @@
 ﻿
 #pragma once
 
-#include "geometry/epoch.hpp"
+#include "astronomy/epoch.hpp"
 
 namespace principia {
-
-using quantities::si::Day;
-
-namespace geometry {
+namespace astronomy {
 
 Instant const jd0  = J2000 - 2451545.0 * Day;
 Instant const mjd0 = J2000 - 51544.5 * Day;
