+++ geometry/geometry.vcxproj
@@ -284,7 +284,6 @@
     <ClInclude Include="affine_map_body.hpp" />
     <ClInclude Include="barycentre_calculator.hpp" />
     <ClInclude Include="barycentre_calculator_body.hpp" />
-    <ClInclude Include="epoch.hpp" />
     <ClInclude Include="epoch_body.hpp" />
     <ClInclude Include="frame.hpp" />
     <ClInclude Include="frame_body.hpp" />
